dms_diffselection scales pseudocount to lower depth sample
Previously, the pseudocounts for dms_diffselection
where always scaled so that the pseudocount was
added to the selected sample. This could give
undesired behavior (e.g., lower effective pseudocount)
when the mock sample actually had less depth. Now
the pseudocount is scaled so that the lower-depth
sample always gets the unscaled pseudocount, so
that the --pseudocount parameter effectively is
the minimum given to either sample at a site.