Fixed bug in ReadDMSCounts / WriteDMSCounts
In a recent commit, *ReadDMSCounts* was changed
to include the frequency and total counts at
each site. This option was undocumented;
documentation was added. In addition, the change
broke the property that *WriteMSCounts* could
write dictionaries read by *ReadDMSCounts*;
this was fixed. Finally, the bug broke the
doctests, so this was also fixed.